requirements.txt
tensorflow
numpy
pandas
matplotlib
pillow
lxml
scikit-learn
streamlit
opendatasets
opencv-python-headless
